chore: extract tsv from facets rds

+++ nextflow/modules/facets.nf
@@ -11,17 +11,23 @@ process FACETS {
 
     output:
     tuple val(meta), path(prefix), emit: cnv
+    tuple val(with_caller), path("${prefix}/${prefix}_hisens.rds"), emit: rds
     tuple val(meta), path("${prefix}/purity.txt"), path("${prefix}/ploidy.txt"), emit: purity_ploidy
+    path(tsv)
     path("*.log")
     //
 
     script:
     prefix = "${module_number}-${meta.filename}-Facets"
     check = file("${meta.out}/${prefix}")
+    tsv = "${prefix}/${prefix}_hisens.tsv"
+    check2 = file("${meta.out}/${tsv}")
+    with_caller = meta + ["caller": "facets"]
     args = task.ext.args.join(" ")
-    if (check.exists()) {
+    if (check.exists() && check2.exists()) {
         """
         cp -r $check .
+        ln -sr ${check2} .
         ln -sr ${meta.log}/facets.log .
         """
     } else {
@@ -34,6 +40,8 @@ process FACETS {
 
         cut -f 2 ${prefix}/${prefix}.txt | tail -n 1 > ${prefix}/purity.txt
         cut -f 3 ${prefix}/${prefix}.txt | tail -n 1 > ${prefix}/ploidy.txt
+
+        get_facets.bash "${prefix}/${prefix}_hisens.rds" "${tsv}"
 
         get_nextflow_log.bash facets.log
         """
